+++ src/main/java/com/ingesup/labojava/controller/HomeController.java
@@ -8,15 +8,15 @@ import org.springframework.web.bind.annotation.RequestMethod;
  * Handles requests for the application home page.
  */
 @Controller
-@RequestMapping(value="/")
 public class HomeController {
 	
-	@RequestMapping(method = RequestMethod.GET)
+
+	@RequestMapping(value="/", method = RequestMethod.GET)
 	public String homePage() {
 		return "home";
 	}
 	
-	@RequestMapping(value="/accueil", method = RequestMethod.GET)
+	@RequestMapping(value="/home", method = RequestMethod.GET)
 	public String indexPage() {
 		return "home";
 	}
