+++ src/main/java/com/ingesup/labojava/controller/ProfileController.java
@@ -1,5 +1,17 @@
 package com.ingesup.labojava.controller;
 
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+@Controller
 public class ProfileController {
+	
+	@RequestMapping(value="/profile", method=RequestMethod.GET)
+	public String profilePage(final Model model) {
+		
+		return "profile";
+	}
 
 }
